requirements.txt
PyOpenGL==3.1.6
